Centralising functionality required by Explode
