+++ Dynamo20/Dynamo_UI/Views/Engine/Explode.cs
@@ -38,74 +38,6 @@ namespace BH.UI.Dynamo.Views
 {
     public class ExplodeView : CallerView<ExplodeComponent> 
     {
-        /*******************************************/
-        /**** Interface Methods                 ****/
-        /*******************************************/
-
-        public override void CustomizeView(ExplodeComponent component, NodeView nodeView)
-        {
-            base.CustomizeView(component, nodeView);
-
-            component.PortConnected += ExplodeComponent_PortConnected;
-            component.PortDisconnected += ExplodeComponent_PortDisconnected;
-        }
-
-
-        /*******************************************/
-        /**** Private Methods                   ****/
-        /*******************************************/
-
-        private void ExplodeComponent_PortDisconnected(PortModel obj)
-        {
-            if (obj.PortType == PortType.Input)
-                UpdateOutputs();
-        }
-
-        /*******************************************/
-
-        private void ExplodeComponent_PortConnected(PortModel arg1, global::Dynamo.Graph.Connectors.ConnectorModel arg2)
-        {
-            if (arg1.PortType == PortType.Input)
-                UpdateOutputs();
-        }
-
-        /*******************************************/
-
-        private void UpdateOutputs()
-        {
-            if (m_DynamoEngine == null || m_Node.InPorts.Count == 0 || m_Node.InPorts[0].Connectors.Count == 0) return;
-
-            NodeModel valuesNode = m_Node.InPorts[0].Connectors[0].Start.Owner;
-            int index = m_Node.InPorts[0].Connectors[0].Start.Index;
-            string startId = valuesNode.GetAstIdentifierForOutputIndex(index).Name;
-            RuntimeMirror colorsMirror = m_DynamoEngine.GetMirror(startId);
-
-            if (colorsMirror == null || colorsMirror.GetData() == null) return;
-
-            MirrorData mirrorData = colorsMirror.GetData();
-            List<object> data = CollectData(mirrorData);
-
-            ExplodeCaller caller = m_Node.Caller as ExplodeCaller;
-            caller.CollectOutputTypes(data);
-        }
-
-        /*******************************************/
-
-        private List<object> CollectData(MirrorData mirrorData)
-        {
-            if (mirrorData.IsCollection)
-            {
-                IEnumerable<MirrorData> elements = mirrorData.GetElements();
-                if (elements.Count() > 0)
-                    return elements.SelectMany(x => CollectData(x)).ToList();
-                else
-                    return new List<object>();
-            }
-            else
-                return new List<object> { mirrorData.Data.IFromDesignScript() };
-        }
-
-        /*******************************************/
     }
 }
